// src/app/model/map.model.ts
export class Mapping {
    id?: any;
    generateur?: string;
    fichier?: string;
    delimiteur?: string;
    lignesAignorer?: string;
    Informatica?: string;
    source?: string;
    destination?: string;
    tablederejet?: string;
    masquefichier?: string;
    masquejava?: string;
    masqueinfo?: string;
    SchemaInput?: string;
    ajouter?:string;

  }